agent/requirements.txt
strands-agents==1.21.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
anthropic==0.40.0
mcp

# Gmail API for attachment download
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
